Improve #includes for topology
