+++ hpx/runtime/threads/topology.hpp
@@ -7,26 +7,36 @@
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 ////////////////////////////////////////////////////////////////////////////////
 
-#if !defined(HPX_E43E0AF0_8A9D_4870_8CC7_E5AD53EF4798)
-#define HPX_E43E0AF0_8A9D_4870_8CC7_E5AD53EF4798
+#ifndef HPX_RUNTIME_THREADS_TOPOLOGY_HPP
+#define HPX_RUNTIME_THREADS_TOPOLOGY_HPP
 
 #include <hpx/config.hpp>
 #include <hpx/exception_fwd.hpp>
 #include <hpx/runtime/naming_fwd.hpp>
 #include <hpx/util/assert.hpp>
 
+#include <boost/cstdint.hpp>
 #include <boost/thread/thread.hpp>
-#include <boost/variant.hpp>
-#include <boost/dynamic_bitset.hpp>
 
+#include <cstddef>
 #include <iosfwd>
 #include <limits>
-#include <string>
-#include <vector>
+#include <utility>
 
 #if defined(HPX_HAVE_MORE_THAN_64_THREADS) || (defined(HPX_HAVE_MAX_CPU_COUNT) \
             && HPX_HAVE_MAX_CPU_COUNT > 64)
+# if defined(HPX_HAVE_MAX_CPU_COUNT)
 #include <bitset>
+# else
+#include <boost/dynamic_bitset.hpp>
+# endif
+#endif
+
+#if defined(HPX_HAVE_HWLOC)
+#include <hpx/error_code.hpp>
+
+#include <string>
+#include <vector>
 #endif
 
 namespace hpx { namespace threads
@@ -386,77 +396,6 @@ namespace hpx { namespace threads
     HPX_API_EXPORT topology const& get_topology();
 
 #if defined(HPX_HAVE_HWLOC)
-    ///////////////////////////////////////////////////////////////////////////
-    namespace detail
-    {
-        typedef std::vector<boost::int64_t> bounds_type;
-
-        enum distribution_type
-        {
-            compact  = 0x01,
-            scatter  = 0x02,
-            balanced = 0x04
-        };
-
-        struct spec_type
-        {
-            enum type { unknown, thread, socket, numanode, core, pu };
-            HPX_API_EXPORT static char const* type_name(type t);
-
-            static boost::int64_t all_entities()
-            {
-                return (std::numeric_limits<boost::int64_t>::min)();
-            }
-
-            spec_type(type t = unknown, boost::int64_t min = all_entities(),
-                    boost::int64_t max = all_entities())
-              : type_(t), index_bounds_()
-            {
-                if (t != unknown) {
-                    if (max == 0 || max == all_entities()) {
-                        // one or all entities
-                        index_bounds_.push_back(min);
-                    }
-                    else if (min != all_entities()) {
-                        // all entities between min and -max, or just min,max
-                        HPX_ASSERT(min >= 0);
-                        index_bounds_.push_back(min);
-                        index_bounds_.push_back(max);
-                    }
-                }
-            }
-
-            bool operator==(spec_type const& rhs) const
-            {
-                if (type_ != rhs.type_ || index_bounds_.size() !=
-                    rhs.index_bounds_.size())
-                    return false;
-
-                for (std::size_t i = 0; i < index_bounds_.size(); ++i)
-                {
-                    if (index_bounds_[i] != rhs.index_bounds_[i])
-                        return false;
-                }
-
-                return true;
-            }
-
-            type type_;
-            bounds_type index_bounds_;
-        };
-
-        typedef std::vector<spec_type> mapping_type;
-        typedef std::pair<spec_type, mapping_type> full_mapping_type;
-        typedef std::vector<full_mapping_type> mappings_spec_type;
-        typedef boost::variant<distribution_type, mappings_spec_type> mappings_type;
-
-        HPX_API_EXPORT bounds_type extract_bounds(spec_type& m,
-            std::size_t default_last, error_code& ec);
-
-        HPX_API_EXPORT void parse_mappings(std::string const& spec,
-            mappings_type& mappings, error_code& ec = throws);
-    }
-
     HPX_API_EXPORT void parse_affinity_options(std::string const& spec,
         std::vector<mask_type>& affinities,
         std::size_t used_cores,
@@ -477,5 +416,4 @@ namespace hpx { namespace threads
     /// \endcond
 }}
 
-#endif // HPX_E43E0AF0_8A9D_4870_8CC7_E5AD53EF4798
-
+#endif /*HPX_RUNTIME_THREADS_TOPOLOGY_HPP*/
