+++ src/runtime/threads/topology.cpp
@@ -4,11 +4,15 @@
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
-#include <hpx/config.hpp>
 #include <hpx/runtime/threads/topology.hpp>
+
+#include <hpx/error_code.hpp>
+#include <hpx/throw_exception.hpp>
 #include <hpx/runtime/threads/policies/topology.hpp>
 #include <hpx/runtime.hpp>
 #include <hpx/util/static.hpp>
+
+#include <cstddef>
 
 #if defined(__ANDROID__) && defined(ANDROID)
 #include <cpu-features.h>
